Move views to separate  directory.
